feat: single depGraph with subject
This commit adds functionality to the sbom service, allowing users to
pass in a single depGraph *and* a subject that will be honored.

To achieve this, we simply send the multi-depgraph body with only one
depGraph, and the subject with it.

Additionally, I've switched from using raw strings to construct the
body-JSON to using `json.RawMessage` and proper types / structs. This
also means that I've had to make the `Name` and `Version` fields of the
`subject` type public. And to make the subject's usage easier in
external packages, I've made the `Subject` type itself public as well.

`json.RawMessage` is nice because it can be used to tell Go that a
certain field is already raw JSON, and has never been fully decoded into
a custom type. As such, when marshaling JSON, that `RawMessage` is
simply added to the rest of the document without any changes.