test: more testing for SBOM
This commit adds integration tests for the opensource SBOM command, as
well as a test package for external users to do the same.

The integration test for the SBOM command uses a mocked legacy CLI and a
mocked SBOM conversion service, but ensures that the `sbom` and
`depgraph` commands work together appropriately.

The mocks have been moved to the `test` package as well, as it was the
most natural place to put it (keep in mind that the SBOM conversion
service mock will have external users as well).