refactor: abstract depgraph opensource implementation
This commit abstracts away the opensource-specific parts of the depgraph
subcommand. The opensource-specific parts are now in
`internal/opensource/depgraph`, while the generic depgraph package stays
at `pkg/depgraph` so that it can be used by external modules as well.

Additionally, it adds a top-level package that exposes that internal
`depgraph` subcommand so that it can be registered at the CLI engine.